requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
python-dotenv==0.21.0
